Solver.py v5
This update I created the guesser class for iterative improvement. I added class functions to see what numbers were needed for any given row. I added the ability to see if which square a given index is in. I also moved the solvable function down to the Matrix class and made it more accurate.