Solver.py v7
I added the solve function to the Matrix Class. Yes I could optimize it by using np arrays from the start but oh well. I also added a bit to the end of iterative improvement, so that If the matrix gets solved in a way that doesn't make any sense, It'll back track and keep working. This is my last push on this file before I start debugging the iterative improvement algorithm.